# client/assets/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bundlephobia API",
    "description": "Read-only APIs for the cost of adding an npm package to a frontend bundle. Sizes are produced by building the package with a real bundler, so a first request for an uncached package version may take several seconds while the build runs.\n\nNo authentication. Requests are rate limited per IP, and requests that trigger a fresh build are limited more tightly than requests served from cache.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "externalDocs": {
      "url": "https://github.com/pastelsky/bundlephobia#readme"
    }
  },
  "servers": [
    {
      "url": "https://bundlephobia.com"
    }
  ],
...
  "paths": {
...
    "/api/exports": {
      "get": {
        "operationId": "getPackageExports",
        "tags": ["size"],
        "summary": "Named exports of an npm package",
        "parameters": [
          { "$ref": "#/components/parameters/package" },
          { "$ref": "#/components/parameters/entryPoint" }
        ],
        "responses": {
          "200": {
            "description": "Map of export name to the resolved file that provides it.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["name", "version", "exports"],
                  "properties": {
                    "name": { "type": "string" },
                    "version": { "type": "string" },
                    "entryPoint": { "type": "string" },
                    "exports": {
                      "type": "object",
                      "additionalProperties": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "422": { "$ref": "#/components/responses/BuildError" }
        }
      }
    },
...
    "/api/entry-points": {
      "get": {
        "operationId": "getPackageEntryPoints",
        "tags": ["discovery"],
        "summary": "List concrete public browser entry points without compiling bundles",
        "parameters": [{ "$ref": "#/components/parameters/package" }],
        "responses": {
          "200": {
            "description": "Resolved package version and buildable import paths; the root may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["name", "version", "entryPoints"],
                  "properties": {
                    "name": { "type": "string" },
                    "version": { "type": "string" },
                    "entryPoints": {
                      "type": "array",
                      "items": { "type": "string" },
                      "examples": [[".", "./map.js"]]
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid query." },
          "404": { "description": "Package or version does not exist." },
          "422": { "$ref": "#/components/responses/BuildError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "entryPoint": {
        "name": "entryPoint",
        "in": "query",
        "description": "Public import path returned by /api/entry-points, such as ./map.js. Omitted or . selects the root. Subpath history contains only matching cached measurements; it does not trigger builds.",
        "schema": { "type": "string", "maxLength": 1024 },
        "example": "./map.js"
      },
      "package": {
        "name": "package",
        "in": "query",
        "required": true,
        "description": "Package name, optionally with a semver version or range, for example `react`, `react@18.2.0` or `@tanstack/react-query`.",
        "schema": { "type": "string" },
        "examples": {
          "unscoped": { "value": "react" },
          "pinned": { "value": "react@18.2.0" },
          "scoped": { "value": "@tanstack/react-query" }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Error class, for example `PackageNotFoundError` or `BuildError`."
          },
          "message": { "type": "string" }
        }
      }
    },
    "responses": {
      "BuildError": {
        "description": "The package could not be resolved or built.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests that require a fresh build. Retry later, or request a package version that is already cached."
      }
    }
  }
}